# perf: Apply filter condition on join to reduce number of rows scanned