Filtering in the list view for  a DocType "Normalised Data Format" (~1k records) by a child table "Normalised Data Point"'s  (~31k records) column "Remark" (is, set) generates the following SQL: 

```sql
SELECT
    `tabNormalised Data Format`.`name`,
    `tabNormalised Data Format`.`owner`,
    `tabNormalised Data Format`.`creation`,
    `tabNormalised Data Format`.`modified`,
    `tabNormalised Data Format`.`modified_by`,
    `tabNormalised Data Format`.`_user_tags`,
    `tabNormalised Data Format`.`_comments`,
    `tabNormalised Data Format`.`_assign`,
    `tabNormalised Data Format`.`_liked_by`,
    `tabNormalised Data Format`.`docstatus`,
    `tabNormalised Data Format`.`idx`,
    `tabNormalised Data Format`.`generated_by`,
    `tabNormalised Data Format`.`is_child` 
FROM
    `tabNormalised Data Format` 
LEFT JOIN
    `tabNormalised Data Point` 
        ON (
            `tabNormalised Data Point`.parenttype = 'Normalised Data Format' 
            AND `tabNormalised Data Point`.parent = `tabNormalised Data Format`.name
        ) 
WHERE
    coalesce(`tabNormalised Data Point`.`remark`, '') != '' 
GROUP BY
    `tabNormalised Data Format`.`name` 
ORDER BY
    `tabNormalised Data Format`.`modified` DESC LIMIT 20 OFFSET 0
```

This causes the join of the whole table followed by a filtration. MariaDB console won't even let you execute this query without enabling a flag LOL. However, if we first filter the table and then join, a lot of effort is saved. The proposed changes generate the following query:

```diff
SELECT
    `tabNormalised Data Format`.`name`,
    `tabNormalised Data Format`.`owner`,
    `tabNormalised Data Format`.`creation`,
    `tabNormalised Data Format`.`modified`,
    `tabNormalised Data Format`.`modified_by`,
    `tabNormalised Data Format`.`_user_tags`,
    `tabNormalised Data Format`.`_comments`,
    `tabNormalised Data Format`.`_assign`,
    `tabNormalised Data Format`.`_liked_by`,
    `tabNormalised Data Format`.`docstatus`,
    `tabNormalised Data Format`.`idx`,
    `tabNormalised Data Format`.`generated_by`,
    `tabNormalised Data Format`.`is_child` 
FROM
    `tabNormalised Data Format` 
LEFT JOIN
    `tabNormalised Data Point` 
        ON (
            `tabNormalised Data Point`.parenttype = 'Normalised Data Format' 
            AND `tabNormalised Data Point`.parent = `tabNormalised Data Format`.name
+          AND coalesce(`tabNormalised Data Point`.`remark`, '') != '' 
        ) 
WHERE
    coalesce(`tabNormalised Data Point`.`remark`, '') != '' 
GROUP BY
    `tabNormalised Data Format`.`name` 
ORDER BY
    `tabNormalised Data Format`.`modified` DESC LIMIT 20 OFFSET 0
```
 
From the `reportview.get` API taking **7s** down to **170ms** (db_query time from 6.5s to 60ms) . This translates to a 1000x improvement (?) while filtering list views by child table values.