@ankush 2023-07-26
if you still have DB can you run this and share `analyze` output of all 3? 

```
SELECT
    `tabNormalised Data Format`.`name`,
    `tabNormalised Data Format`.`owner`,
    `tabNormalised Data Format`.`creation`,
    `tabNormalised Data Format`.`modified`,
    `tabNormalised Data Format`.`modified_by`,
    `tabNormalised Data Format`.`_user_tags`,
    `tabNormalised Data Format`.`_comments`,
    `tabNormalised Data Format`.`_assign`,
    `tabNormalised Data Format`.`_liked_by`,
    `tabNormalised Data Format`.`docstatus`,
    `tabNormalised Data Format`.`idx`,
    `tabNormalised Data Format`.`generated_by`,
    `tabNormalised Data Format`.`is_child` 
FROM
    `tabNormalised Data Format` 
LEFT JOIN
    `tabNormalised Data Point` 
        ON (
            `tabNormalised Data Point`.parenttype = 'Normalised Data Format' 
            AND `tabNormalised Data Point`.parent = `tabNormalised Data Format`.name
        ) 
WHERE
    `tabNormalised Data Point`.`remark` != '' 
GROUP BY
    `tabNormalised Data Format`.`name` 
ORDER BY
    `tabNormalised Data Format`.`modified` DESC LIMIT 20 OFFSET 0
```