@gavindsouza 2023-07-26
Won't null != null and null == null will always be Falsy? is (not) Null is the only way that I know of to compare Nulls. 

How is coalesce pointless? If a value has never been set, it stays as NULL. If a value has been set and later unset, Frappe makes it ''. Coalesce makes sure to return '' when it comes across null which we check for equality again with ''.

Can you share why you think the query planner is being confused?

---
I'll analyze the query and share the results tomorrow