@ankush 2023-07-26
| value type | is it "set"? | coelsce(value, '') != '' | value != '' |
| --- | --- | --- | --- |
| `'some string or truth value'` | yes | true | true
| `''` (empty string) | no | false | false |
| `null` | no | false | null |