@ankush 2023-07-26
> Can you share why you think the query planner is being confused?

Because it's a not a deterministic function, so MySQL can't just assume that what you want to filter on is just a column and not some computed value on top of a column.