@gavindsouza 2023-08-03
Still facing a part of the problem on my site but I don't have the bandwidth to look into this now.